change common file and widget testing
